# swagger.yaml
swagger: '2.0'
info:
  title: Bank Argent API documentation
  description: Contains all available API endpoints in this codebase
  version: '1.0.0'
  termsOfService: 'http://swagger.io/terms/'
host: localhost:3001
basePath: /api/v1
schemes:
  - http
paths:
  /user/login:
    post:
      tags:
        - User Module
      summary: Login
      description: API for Login
      parameters:
        - in: body
          name: body
          description: Login Payload
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      produces:
        - application/json
      responses:
        '200':
          description: Login Successfully
          schema:
            $ref: '#/definitions/LoginResponse'
        '400':
          description: Invalid Fields
        '500':
          description: Internal Server Error
  /user/signup:
    post:
      tags:
        - User Module
      summary: Signup
      description: API for Signup
      parameters:
        - in: body
          name: body
          description: Signup Payload
          required: true
          schema:
            $ref: '#/definitions/SignupRequest'
      produces:
        - application/json
      responses:
        '200':
          description: Signup Successfully
          schema:
            $ref: '#/definitions/ApiResponse'
        '400':
          description: Invalid Fields
        '500':
          description: Internal Server Error
  /user/logout:
    delete:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: Logout
      description: Invalidate the auth token
      responses:
        '200':
          description: Logout successfuly
          schema:
            $ref: '#/definitions/ApiResponse'
        '400':
          description: Missing token
        '500':
          description: Internal server error
  /user/profile:
    get:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: User Profile API
      description: API for fetching a user profile
      responses:
        '200':
          description: User profile retrieved successully
          schema:
            $ref: '#/definitions/UserProfileResponse'
        '400':
          description: Invalid Fields
        '401':
          description: Not authorized
        '500':
          description: Internal Server Error
    put:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: User Profile API
      description: API for updating a user profile
      parameters:
        - in: body
          name: body
          description: Update user profile attributes
          required: true
          schema:
            $ref: '#/definitions/PutUserProfileRequest'
      produces:
        - application/json
      responses:
        '200':
          description: User profile retrieved successully
          schema:
            $ref: '#/definitions/PutUserProfileResponse'
        '400':
          description: Invalid Fields
        '401':
          description: Not authorized
        '500':
          description: Internal Server Error
  /transactions/account:
    get:
      security:
        - Bearer: []
      tags:
        - Transactions
      summary: Accounts of the user
      description: Fetch the accounts of the connected user
      responses:
        "200":
          description: Return an array of all accounts
          schema:
            $ref: "#/definitions/TransactionsAccount"
        "500":
          description: Internal Server Error
  /transactions/account/{accountId}:
    get:
      security:
        - Bearer: []
      tags:
        - Transactions
      summary: Transactions By account
      description: Fetch the transactions for the actual month by account
      parameters:
        - name: accountId
          in: path
          description: ID of account to return
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Return an array of transaction according to the account
          schema:
            $ref: "#/definitions/transactionByAccount"
        "404":
          description: Account not found
        "500":
          description: Internal Server Error
  /transactions/{transactionId}:
    get:
      security:
        - Bearer: []
      tags:
        - Transactions
      summary: Transaction details
      description: Fetch the details of the transaction
      parameters:
        - name: transactionId
          in: path
          description: ID of the transaction to return
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Return an object with the details
          schema:
            $ref: "#/definitions/transactionDetails"
        "404":
          description: Transaction not found
        "500":
          description: Internal Server Error
    put:
      security:
        - Bearer: []
      tags:
        - Transactions
      summary: Update transactions details
      description: Update the details of the transaction by id
      parameters:
        - name: transactionId
          in: path
          description: ID of the transaction to update
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Update the details of the transaction
          required: true
          schema:
            $ref: "#/definitions/transactionDetails"
      responses:
        "200":
          description: Update with success
        "400":
          description: Invalid fields
        "404":
          description: Transaction not found
        "500":
          description: Internal Server Error
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
  User:
    type: object
    properties:
      email:
        type: string
        description: user email
      password:
        type: string
        description: user password
      firstName:
        type: string
        description: user first name
      lastName:
        type: string
        description: user last name
      userName:
        type: string
        description: user public name
  ApiResponse:
    type: object
    properties:
      status:
        type: integer
      message:
        type: string
  LoginRequest:
    properties:
      email:
        type: string
        description: user email
      password:
        type: string
        description: user password
  LoginResponse:
    allOf:
      - $ref: "#/definitions/ApiResponse"
      - type: object
        properties:
          body:
            type: object
            properties:
              token:
               type: string
  SignupRequest:
    allOf:
      - $ref: "#/definitions/User"
  UserProfileResponse:
    allOf:
      - $ref: "#/definitions/ApiResponse"
      - type: object
        properties:
          body:
            type: object
            properties:
              id:
                type: string
              email:
                type: string
              firstName:
                type: string
              lastName:
                type: string
              userName:
                type: string
  PutUserProfileRequest:
    type: object
    properties:
      userName:
        type: string
  PutUserProfileResponse:
    allOf:
      - $ref: "#/definitions/ApiResponse"
      - type: object
        properties:
          body:
            type: object
            properties:
              userName:
                type: string
  TransactionsAccount:
    type: array
    items: 
      type: object
      properties:
        accountId:
          type: integer
        name:
          type: string
        amount: 
          type: integer
  transactionByAccount:
    type: array
    items:
      type: object
      properties:
        transactionId:
          type: integer
          format: int64
          example: 123456
        accountId:
          type: integer
          format: int64
          example: 654321
        date:
          type: string
          example: "2024-11-26T13:19:42.568Z"
        transactionDescription:
          type: string
          example: "Golden Sun Bakery"
        amountInDollar:
          type: integer
          format: int64
          example: 8.2
  transactionDetails:
    type: object
    properties:
      transactionId:
        type: integer
        format: int64
        example: 156789
      transactionType:
        type: string
        example: "Electronic"
      category:
        type: string
        example: "Food"
      note:
        type: string
        example: "lorem ipsum"